refactor: Mail 및 재학생 인증 관련 클래스들의 패키지를 목적에 맞게 정리
