--- src/main/java/com/gdschongik/gdsc/domain/integration/UnivMailVerificationConstant.java
@@ -1,14 +1,0 @@
-package com.gdschongik.gdsc.domain.integration;
-
-public class UnivMailVerificationConstant {
-
-    public static final String HONGIK_UNIV_MAIL_DOMAIN = "@g.hongik.ac.kr";
-
-    public static final String VERIFY_EMAIL_ENDPOINT = "/onboarding/verify-email";
-
-    public static final String VERIFY_EMAIL_QUERY_STRING_KEY = "?token=";
-
-    public static final String VERIFICATION_EMAIL_SUBJECT = "GDSC Hongik 이메일 인증 코드입니다.";
-
-    private UnivMailVerificationConstant() {}
-}
