+++ src/main/java/com/gdschongik/gdsc/domain/integration/application/UnivEmailVerificationService.java
@@ -1,5 +1,6 @@
-package com.gdschongik.gdsc.domain.integration;
+package com.gdschongik.gdsc.domain.integration.application;
 
+import com.gdschongik.gdsc.domain.integration.dao.VerificationCodeAndEmailRepository;
 import com.gdschongik.gdsc.domain.member.dao.MemberRepository;
 import com.gdschongik.gdsc.domain.member.domain.Member;
 import com.gdschongik.gdsc.domain.member.domain.RequirementStatus;
