+++ src/main/java/com/gdschongik/gdsc/domain/integration/dao/VerificationCodeAndEmailRepository.java
@@ -1,5 +1,6 @@
-package com.gdschongik.gdsc.domain.integration;
+package com.gdschongik.gdsc.domain.integration.dao;
 
+import com.gdschongik.gdsc.domain.integration.domain.VerificationCodeAndEmail;
 import org.springframework.data.repository.CrudRepository;
 
 public interface VerificationCodeAndEmailRepository extends CrudRepository<VerificationCodeAndEmail, String> {}
