+++ src/main/java/com/gdschongik/gdsc/global/common/constant/EmailConstant.java
@@ -6,5 +6,7 @@ public class EmailConstant {
 
     public static final String SENDER_ADDRESS = "gdsc.hongik@gmail.com";
 
+    public static final String HONGIK_UNIV_MAIL_DOMAIN = "@g.hongik.ac.kr";
+
     private EmailConstant() {}
 }
